requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
django-cors-headers==1.0.0
requests==2.3.0
django-nose==1.4.3
coverage
